Move features into a separate module out of msgs.
