+++ lightning/src/ln/functional_tests.rs
@@ -11,8 +11,9 @@ use ln::channelmonitor::{ChannelMonitor, CLTV_CLAIM_BUFFER, LATENCY_GRACE_PERIOD
 use ln::channel::{ACCEPTED_HTLC_SCRIPT_WEIGHT, OFFERED_HTLC_SCRIPT_WEIGHT, Channel, ChannelError};
 use ln::onion_utils;
 use ln::router::{Route, RouteHop};
+use ln::features::{ChannelFeatures, InitFeatures};
 use ln::msgs;
-use ln::msgs::{ChannelMessageHandler,RoutingMessageHandler,HTLCFailChannelUpdate, Features, FeatureContextChannel, FeatureContextInit, ErrorAction};
+use ln::msgs::{ChannelMessageHandler,RoutingMessageHandler,HTLCFailChannelUpdate, ErrorAction};
 use util::enforcing_trait_impls::EnforcingChannelKeys;
 use util::test_utils;
 use util::events::{Event, EventsProvider, MessageSendEvent, MessageSendEventsProvider};
@@ -69,7 +70,7 @@ fn test_insane_channel_opens() {
 	// Test helper that asserts we get the correct error string given a mutator
 	// that supposedly makes the channel open message insane
 	let insane_open_helper = |expected_error_str: &str, message_mutator: fn(msgs::OpenChannel) -> msgs::OpenChannel| {
-		nodes[1].node.handle_open_channel(&nodes[0].node.get_our_node_id(), Features::<FeatureContextInit>::supported(), &message_mutator(open_channel_message.clone()));
+		nodes[1].node.handle_open_channel(&nodes[0].node.get_our_node_id(), InitFeatures::supported(), &message_mutator(open_channel_message.clone()));
 		let msg_events = nodes[1].node.get_and_clear_pending_msg_events();
 		assert_eq!(msg_events.len(), 1);
 		if let MessageSendEvent::HandleError { ref action, .. } = msg_events[0] {
@@ -108,7 +109,7 @@ fn test_insane_channel_opens() {
 #[test]
 fn test_async_inbound_update_fee() {
 	let mut nodes = create_network(2, &[None, None]);
-	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 	let channel_id = chan.2;
 
 	// balancing
@@ -218,7 +219,7 @@ fn test_update_fee_unordered_raa() {
 	// Just the intro to the previous test followed by an out-of-order RAA (which caused a
 	// crash in an earlier version of the update_fee patch)
 	let mut nodes = create_network(2, &[None, None]);
-	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 	let channel_id = chan.2;
 
 	// balancing
@@ -268,7 +269,7 @@ fn test_update_fee_unordered_raa() {
 #[test]
 fn test_multi_flight_update_fee() {
 	let nodes = create_network(2, &[None, None]);
-	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 	let channel_id = chan.2;
 
 	// A                                        B
@@ -372,7 +373,7 @@ fn test_multi_flight_update_fee() {
 #[test]
 fn test_update_fee_vanilla() {
 	let nodes = create_network(2, &[None, None]);
-	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 	let channel_id = chan.2;
 
 	let feerate = get_feerate!(nodes[0], channel_id);
@@ -411,7 +412,7 @@ fn test_update_fee_vanilla() {
 fn test_update_fee_that_funder_cannot_afford() {
 	let nodes = create_network(2, &[None, None]);
 	let channel_value = 1888;
-	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, channel_value, 700000, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, channel_value, 700000, InitFeatures::supported(), InitFeatures::supported());
 	let channel_id = chan.2;
 
 	let feerate = 260;
@@ -459,7 +460,7 @@ fn test_update_fee_that_funder_cannot_afford() {
 #[test]
 fn test_update_fee_with_fundee_update_add_htlc() {
 	let mut nodes = create_network(2, &[None, None]);
-	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 	let channel_id = chan.2;
 
 	// balancing
@@ -553,7 +554,7 @@ fn test_update_fee_with_fundee_update_add_htlc() {
 #[test]
 fn test_update_fee() {
 	let nodes = create_network(2, &[None, None]);
-	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 	let channel_id = chan.2;
 
 	// A                                        B
@@ -654,7 +655,7 @@ fn test_update_fee() {
 fn pre_funding_lock_shutdown_test() {
 	// Test sending a shutdown prior to funding_locked after funding generation
 	let nodes = create_network(2, &[None, None]);
-	let tx = create_chan_between_nodes_with_value_init(&nodes[0], &nodes[1], 8000000, 0, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let tx = create_chan_between_nodes_with_value_init(&nodes[0], &nodes[1], 8000000, 0, InitFeatures::supported(), InitFeatures::supported());
 	let header = BlockHeader { version: 0x20000000, prev_blockhash: Default::default(), merkle_root: Default::default(), time: 42, bits: 42, nonce: 42 };
 	nodes[0].block_notifier.block_connected(&Block { header, txdata: vec![tx.clone()]}, 1);
 	nodes[1].block_notifier.block_connected(&Block { header, txdata: vec![tx.clone()]}, 1);
@@ -680,8 +681,8 @@ fn pre_funding_lock_shutdown_test() {
 fn updates_shutdown_wait() {
 	// Test sending a shutdown with outstanding updates pending
 	let mut nodes = create_network(3, &[None, None, None]);
-	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
+	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, InitFeatures::supported(), InitFeatures::supported());
 	let route_1 = nodes[0].router.get_route(&nodes[1].node.get_our_node_id(), None, &[], 100000, TEST_FINAL_CLTV).unwrap();
 	let route_2 = nodes[1].router.get_route(&nodes[0].node.get_our_node_id(), None, &[], 100000, TEST_FINAL_CLTV).unwrap();
 
@@ -752,8 +753,8 @@ fn updates_shutdown_wait() {
 fn htlc_fail_async_shutdown() {
 	// Test HTLCs fail if shutdown starts even if messages are delivered out-of-order
 	let mut nodes = create_network(3, &[None, None, None]);
-	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
+	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, InitFeatures::supported(), InitFeatures::supported());
 
 	let route = nodes[0].router.get_route(&nodes[2].node.get_our_node_id(), None, &[], 100000, TEST_FINAL_CLTV).unwrap();
 	let (_, our_payment_hash) = get_payment_preimage_hash!(nodes[0]);
@@ -833,8 +834,8 @@ fn do_test_shutdown_rebroadcast(recv_count: u8) {
 	// Test that shutdown/closing_signed is re-sent on reconnect with a variable number of
 	// messages delivered prior to disconnect
 	let nodes = create_network(3, &[None, None, None]);
-	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
+	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, InitFeatures::supported(), InitFeatures::supported());
 
 	let (our_payment_preimage, _) = route_payment(&nodes[0], &[&nodes[1], &nodes[2]], 100000);
 
@@ -994,9 +995,9 @@ fn fake_network_test() {
 	let nodes = create_network(4, &[None, None, None, None]);
 
 	// Create some initial channels
-	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	let chan_3 = create_announced_chan_between_nodes(&nodes, 2, 3, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
+	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, InitFeatures::supported(), InitFeatures::supported());
+	let chan_3 = create_announced_chan_between_nodes(&nodes, 2, 3, InitFeatures::supported(), InitFeatures::supported());
 
 	// Rebalance the network a bit by relaying one payment through all the channels...
 	send_payment(&nodes[0], &vec!(&nodes[1], &nodes[2], &nodes[3])[..], 8000000, 8_000_000);
@@ -1014,7 +1015,7 @@ fn fake_network_test() {
 	fail_payment(&nodes[0], &vec!(&nodes[1], &nodes[2], &nodes[3])[..], payment_hash_1);
 
 	// Add a new channel that skips 3
-	let chan_4 = create_announced_chan_between_nodes(&nodes, 1, 3, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_4 = create_announced_chan_between_nodes(&nodes, 1, 3, InitFeatures::supported(), InitFeatures::supported());
 
 	send_payment(&nodes[0], &vec!(&nodes[1], &nodes[3])[..], 1000000, 1_000_000);
 	send_payment(&nodes[2], &vec!(&nodes[3])[..], 1000000, 1_000_000);
@@ -1076,7 +1077,7 @@ fn fake_network_test() {
 	claim_payment(&nodes[1], &vec!(&nodes[2], &nodes[3], &nodes[1])[..], payment_preimage_1, 1_000_000);
 
 	// Add a duplicate new channel from 2 to 4
-	let chan_5 = create_announced_chan_between_nodes(&nodes, 1, 3, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_5 = create_announced_chan_between_nodes(&nodes, 1, 3, InitFeatures::supported(), InitFeatures::supported());
 
 	// Send some payments across both channels
 	let payment_preimage_3 = route_payment(&nodes[0], &vec!(&nodes[1], &nodes[3])[..], 3000000).0;
@@ -1109,8 +1110,8 @@ fn holding_cell_htlc_counting() {
 	// to ensure we don't end up with HTLCs sitting around in our holding cell for several
 	// commitment dance rounds.
 	let mut nodes = create_network(3, &[None, None, None]);
-	create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
+	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, InitFeatures::supported(), InitFeatures::supported());
 
 	let mut payments = Vec::new();
 	for _ in 0..::ln::channel::OUR_MAX_HTLCS {
@@ -1238,11 +1239,11 @@ fn duplicate_htlc_test() {
 	let mut nodes = create_network(6, &[None, None, None, None, None, None]);
 
 	// Create some initial channels to route via 3 to 4/5 from 0/1/2
-	create_announced_chan_between_nodes(&nodes, 0, 3, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	create_announced_chan_between_nodes(&nodes, 1, 3, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	create_announced_chan_between_nodes(&nodes, 2, 3, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	create_announced_chan_between_nodes(&nodes, 3, 4, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	create_announced_chan_between_nodes(&nodes, 3, 5, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	create_announced_chan_between_nodes(&nodes, 0, 3, InitFeatures::supported(), InitFeatures::supported());
+	create_announced_chan_between_nodes(&nodes, 1, 3, InitFeatures::supported(), InitFeatures::supported());
+	create_announced_chan_between_nodes(&nodes, 2, 3, InitFeatures::supported(), InitFeatures::supported());
+	create_announced_chan_between_nodes(&nodes, 3, 4, InitFeatures::supported(), InitFeatures::supported());
+	create_announced_chan_between_nodes(&nodes, 3, 5, InitFeatures::supported(), InitFeatures::supported());
 
 	let (payment_preimage, payment_hash) = route_payment(&nodes[0], &vec!(&nodes[3], &nodes[4])[..], 1000000);
 
@@ -1264,7 +1265,7 @@ fn test_duplicate_htlc_different_direction_onchain() {
 	// in opposite directions.
 	let nodes = create_network(2, &[None, None]);
 
-	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	// balancing
 	send_payment(&nodes[0], &vec!(&nodes[1])[..], 8000000, 8_000_000);
@@ -1330,8 +1331,8 @@ fn test_duplicate_htlc_different_direction_onchain() {
 fn do_channel_reserve_test(test_recv: bool) {
 
 	let mut nodes = create_network(3, &[None, None, None]);
-	let chan_1 = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1900, 1001, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	let chan_2 = create_announced_chan_between_nodes_with_value(&nodes, 1, 2, 1900, 1001, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_1 = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1900, 1001, InitFeatures::supported(), InitFeatures::supported());
+	let chan_2 = create_announced_chan_between_nodes_with_value(&nodes, 1, 2, 1900, 1001, InitFeatures::supported(), InitFeatures::supported());
 
 	let mut stat01 = get_channel_value_stat!(nodes[0], chan_1.2);
 	let mut stat11 = get_channel_value_stat!(nodes[1], chan_1.2);
@@ -1614,7 +1615,7 @@ fn channel_reserve_in_flight_removes() {
 	//  * Now B happily sends another HTLC, potentially violating its reserve value from A's point
 	//    of view (if A counts the AwaitingRemovedRemoteRevoke HTLC).
 	let mut nodes = create_network(2, &[None, None]);
-	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	let b_chan_values = get_channel_value_stat!(nodes[1], chan_1.2);
 	// Route the first two HTLCs.
@@ -1743,10 +1744,10 @@ fn channel_monitor_network_test() {
 	let nodes = create_network(5, &[None, None, None, None, None]);
 
 	// Create some initial channels
-	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	let chan_3 = create_announced_chan_between_nodes(&nodes, 2, 3, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	let chan_4 = create_announced_chan_between_nodes(&nodes, 3, 4, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
+	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, InitFeatures::supported(), InitFeatures::supported());
+	let chan_3 = create_announced_chan_between_nodes(&nodes, 2, 3, InitFeatures::supported(), InitFeatures::supported());
+	let chan_4 = create_announced_chan_between_nodes(&nodes, 3, 4, InitFeatures::supported(), InitFeatures::supported());
 
 	// Rebalance the network a bit by relaying one payment through all the channels...
 	send_payment(&nodes[0], &vec!(&nodes[1], &nodes[2], &nodes[3], &nodes[4])[..], 8000000, 8_000_000);
@@ -1890,7 +1891,7 @@ fn test_justice_tx() {
 	let cfgs = [Some(alice_config), Some(bob_config)];
 	let nodes = create_network(2, &cfgs);
 	// Create some new channels:
-	let chan_5 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_5 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	// A pending HTLC which will be revoked:
 	let payment_preimage_3 = route_payment(&nodes[0], &vec!(&nodes[1])[..], 3000000).0;
@@ -1934,7 +1935,7 @@ fn test_justice_tx() {
 
 	// We test justice_tx build by A on B's revoked HTLC-Success tx
 	// Create some new channels:
-	let chan_6 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_6 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 	{
 		let mut node_txn = nodes[1].tx_broadcaster.txn_broadcasted.lock().unwrap();
 		node_txn.clear();
@@ -1980,7 +1981,7 @@ fn revoked_output_claim() {
 	// Simple test to ensure a node will claim a revoked output when a stale remote commitment
 	// transaction is broadcast by its counterparty
 	let nodes = create_network(2, &[None, None]);
-	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 	// node[0] is gonna to revoke an old state thus node[1] should be able to claim the revoked output
 	let revoked_local_txn = nodes[0].node.channel_state.lock().unwrap().by_id.get_mut(&chan_1.2).unwrap().channel_monitor().get_latest_local_commitment_txn();
 	assert_eq!(revoked_local_txn.len(), 1);
@@ -2011,7 +2012,7 @@ fn claim_htlc_outputs_shared_tx() {
 	let nodes = create_network(2, &[None, None]);
 
 	// Create some new channel:
-	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	// Rebalance the network to generate htlc in the two directions
 	send_payment(&nodes[0], &vec!(&nodes[1])[..], 8000000, 8_000_000);
@@ -2085,7 +2086,7 @@ fn claim_htlc_outputs_single_tx() {
 	// Node revoked old state, htlcs have timed out, claim each of them in separated justice tx
 	let nodes = create_network(2, &[None, None]);
 
-	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	// Rebalance the network to generate htlc in the two directions
 	send_payment(&nodes[0], &vec!(&nodes[1])[..], 8000000, 8_000_000);
@@ -2184,8 +2185,8 @@ fn test_htlc_on_chain_success() {
 	let nodes = create_network(3, &[None, None, None]);
 
 	// Create some initial channels
-	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
+	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, InitFeatures::supported(), InitFeatures::supported());
 
 	// Rebalance the network a bit by relaying one payment through all the channels...
 	send_payment(&nodes[0], &vec!(&nodes[1], &nodes[2])[..], 8000000, 8_000_000);
@@ -2349,8 +2350,8 @@ fn test_htlc_on_chain_timeout() {
 	let nodes = create_network(3, &[None, None, None]);
 
 	// Create some intial channels
-	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
+	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, InitFeatures::supported(), InitFeatures::supported());
 
 	// Rebalance the network a bit by relaying one payment thorugh all the channels...
 	send_payment(&nodes[0], &vec!(&nodes[1], &nodes[2])[..], 8000000, 8_000_000);
@@ -2457,8 +2458,8 @@ fn test_simple_commitment_revoked_fail_backward() {
 	let nodes = create_network(3, &[None, None, None]);
 
 	// Create some initial channels
-	create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
+	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, InitFeatures::supported(), InitFeatures::supported());
 
 	let (payment_preimage, _payment_hash) = route_payment(&nodes[0], &[&nodes[1], &nodes[2]], 3000000);
 	// Get the will-be-revoked local txn from nodes[2]
@@ -2525,8 +2526,8 @@ fn do_test_commitment_revoked_fail_backward_exhaustive(deliver_bs_raa: bool, use
 	let mut nodes = create_network(3, &[None, None, None]);
 
 	// Create some initial channels
-	create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
+	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, InitFeatures::supported(), InitFeatures::supported());
 
 	let (payment_preimage, _payment_hash) = route_payment(&nodes[0], &[&nodes[1], &nodes[2]], if no_to_remote { 10_000 } else { 3_000_000 });
 	// Get the will-be-revoked local txn from nodes[2]
@@ -2734,7 +2735,7 @@ fn test_htlc_ignore_latest_remote_commitment() {
 	// Test that HTLC transactions spending the latest remote commitment transaction are simply
 	// ignored if we cannot claim them. This originally tickled an invalid unwrap().
 	let nodes = create_network(2, &[None, None]);
-	create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	route_payment(&nodes[0], &[&nodes[1]], 10000000);
 	nodes[0].node.force_close_channel(&nodes[0].node.list_channels()[0].channel_id);
@@ -2756,8 +2757,8 @@ fn test_htlc_ignore_latest_remote_commitment() {
 fn test_force_close_fail_back() {
 	// Check which HTLCs are failed-backwards on channel force-closure
 	let mut nodes = create_network(3, &[None, None, None]);
-	create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	create_announced_chan_between_nodes(&nodes, 1, 2, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
+	create_announced_chan_between_nodes(&nodes, 1, 2, InitFeatures::supported(), InitFeatures::supported());
 
 	let route = nodes[0].router.get_route(&nodes[2].node.get_our_node_id(), None, &Vec::new(), 1000000, 42).unwrap();
 
@@ -2830,7 +2831,7 @@ fn test_force_close_fail_back() {
 fn test_unconf_chan() {
 	// After creating a chan between nodes, we disconnect all blocks previously seen to force a channel close on nodes[0] side
 	let nodes = create_network(2, &[None, None]);
-	create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	let channel_state = nodes[0].node.channel_state.lock().unwrap();
 	assert_eq!(channel_state.by_id.len(), 1);
@@ -2859,8 +2860,8 @@ fn test_unconf_chan() {
 fn test_simple_peer_disconnect() {
 	// Test that we can reconnect when there are no lost messages
 	let nodes = create_network(3, &[None, None, None]);
-	create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	create_announced_chan_between_nodes(&nodes, 1, 2, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
+	create_announced_chan_between_nodes(&nodes, 1, 2, InitFeatures::supported(), InitFeatures::supported());
 
 	nodes[0].node.peer_disconnected(&nodes[1].node.get_our_node_id(), false);
 	nodes[1].node.peer_disconnected(&nodes[0].node.get_our_node_id(), false);
@@ -2913,10 +2914,10 @@ fn do_test_drop_messages_peer_disconnect(messages_delivered: u8) {
 	// Test that we can reconnect when in-flight HTLC updates get dropped
 	let mut nodes = create_network(2, &[None, None]);
 	if messages_delivered == 0 {
-		create_chan_between_nodes_with_value_a(&nodes[0], &nodes[1], 100000, 10001, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+		create_chan_between_nodes_with_value_a(&nodes[0], &nodes[1], 100000, 10001, InitFeatures::supported(), InitFeatures::supported());
 		// nodes[1] doesn't receive the funding_locked message (it'll be re-sent on reconnect)
 	} else {
-		create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+		create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 	}
 
 	let route = nodes[0].router.get_route(&nodes[1].node.get_our_node_id(), Some(&nodes[0].node.list_usable_channels()), &Vec::new(), 1000000, TEST_FINAL_CLTV).unwrap();
@@ -3118,7 +3119,7 @@ fn test_drop_messages_peer_disconnect_b() {
 fn test_funding_peer_disconnect() {
 	// Test that we can lock in our funding tx while disconnected
 	let nodes = create_network(2, &[None, None]);
-	let tx = create_chan_between_nodes_with_value_init(&nodes[0], &nodes[1], 100000, 10001, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let tx = create_chan_between_nodes_with_value_init(&nodes[0], &nodes[1], 100000, 10001, InitFeatures::supported(), InitFeatures::supported());
 
 	nodes[0].node.peer_disconnected(&nodes[1].node.get_our_node_id(), false);
 	nodes[1].node.peer_disconnected(&nodes[0].node.get_our_node_id(), false);
@@ -3200,7 +3201,7 @@ fn test_drop_messages_peer_disconnect_dual_htlc() {
 	// Test that we can handle reconnecting when both sides of a channel have pending
 	// commitment_updates when we disconnect.
 	let mut nodes = create_network(2, &[None, None]);
-	create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	let (payment_preimage_1, _) = route_payment(&nodes[0], &[&nodes[1]], 1000000);
 
@@ -3340,7 +3341,7 @@ fn test_invalid_channel_announcement() {
 	let secp_ctx = Secp256k1::new();
 	let nodes = create_network(2, &[None, None]);
 
-	let chan_announcement = create_chan_between_nodes(&nodes[0], &nodes[1], Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_announcement = create_chan_between_nodes(&nodes[0], &nodes[1], InitFeatures::supported(), InitFeatures::supported());
 
 	let a_channel_lock = nodes[0].node.channel_state.lock().unwrap();
 	let b_channel_lock = nodes[1].node.channel_state.lock().unwrap();
@@ -3362,7 +3363,7 @@ fn test_invalid_channel_announcement() {
 	macro_rules! dummy_unsigned_msg {
 		() => {
 			msgs::UnsignedChannelAnnouncement {
-				features: msgs::Features::<FeatureContextChannel>::supported(),
+				features: ChannelFeatures::supported(),
 				chain_hash: genesis_block(Network::Testnet).header.bitcoin_hash(),
 				short_channel_id: as_chan.get_short_channel_id().unwrap(),
 				node_id_1: if were_node_one { as_network_key } else { bs_network_key },
@@ -3412,7 +3413,7 @@ fn test_invalid_channel_announcement() {
 fn test_no_txn_manager_serialize_deserialize() {
 	let mut nodes = create_network(2, &[None, None]);
 
-	let tx = create_chan_between_nodes_with_value_init(&nodes[0], &nodes[1], 100000, 10001, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let tx = create_chan_between_nodes_with_value_init(&nodes[0], &nodes[1], 100000, 10001, InitFeatures::supported(), InitFeatures::supported());
 
 	nodes[1].node.peer_disconnected(&nodes[0].node.get_our_node_id(), false);
 
@@ -3474,7 +3475,7 @@ fn test_no_txn_manager_serialize_deserialize() {
 #[test]
 fn test_simple_manager_serialize_deserialize() {
 	let mut nodes = create_network(2, &[None, None]);
-	create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	let (our_payment_preimage, _) = route_payment(&nodes[0], &[&nodes[1]], 1000000);
 	let (_, our_payment_hash) = route_payment(&nodes[0], &[&nodes[1]], 1000000);
@@ -3521,9 +3522,9 @@ fn test_simple_manager_serialize_deserialize() {
 fn test_manager_serialize_deserialize_inconsistent_monitor() {
 	// Test deserializing a ChannelManager with an out-of-date ChannelMonitor
 	let mut nodes = create_network(4, &[None, None, None, None]);
-	create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	create_announced_chan_between_nodes(&nodes, 2, 0, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	let (_, _, channel_id, funding_tx) = create_announced_chan_between_nodes(&nodes, 0, 3, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
+	create_announced_chan_between_nodes(&nodes, 2, 0, InitFeatures::supported(), InitFeatures::supported());
+	let (_, _, channel_id, funding_tx) = create_announced_chan_between_nodes(&nodes, 0, 3, InitFeatures::supported(), InitFeatures::supported());
 
 	let (our_payment_preimage, _) = route_payment(&nodes[2], &[&nodes[0], &nodes[1]], 1000000);
 
@@ -3718,7 +3719,7 @@ fn test_claim_sizeable_push_msat() {
 	// Incidentally test SpendableOutput event generation due to detection of to_local output on commitment tx
 	let nodes = create_network(2, &[None, None]);
 
-	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 100000, 99000000, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 100000, 99000000, InitFeatures::supported(), InitFeatures::supported());
 	nodes[1].node.force_close_channel(&chan.2);
 	check_closed_broadcast!(nodes[1], false);
 	let node_txn = nodes[1].tx_broadcaster.txn_broadcasted.lock().unwrap();
@@ -3739,7 +3740,7 @@ fn test_claim_on_remote_sizeable_push_msat() {
 	// to_remote output is encumbered by a P2WPKH
 	let nodes = create_network(2, &[None, None]);
 
-	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 100000, 99000000, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 100000, 99000000, InitFeatures::supported(), InitFeatures::supported());
 	nodes[0].node.force_close_channel(&chan.2);
 	check_closed_broadcast!(nodes[0], false);
 
@@ -3764,7 +3765,7 @@ fn test_claim_on_remote_revoked_sizeable_push_msat() {
 
 	let nodes = create_network(2, &[None, None]);
 
-	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 100000, 59000000, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 100000, 59000000, InitFeatures::supported(), InitFeatures::supported());
 	let payment_preimage = route_payment(&nodes[0], &vec!(&nodes[1])[..], 3000000).0;
 	let revoked_local_txn = nodes[0].node.channel_state.lock().unwrap().by_id.get_mut(&chan.2).unwrap().channel_monitor().get_latest_local_commitment_txn();
 	assert_eq!(revoked_local_txn[0].input.len(), 1);
@@ -3789,7 +3790,7 @@ fn test_static_spendable_outputs_preimage_tx() {
 	let nodes = create_network(2, &[None, None]);
 
 	// Create some initial channels
-	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	let payment_preimage = route_payment(&nodes[0], &vec!(&nodes[1])[..], 3000000).0;
 
@@ -3833,7 +3834,7 @@ fn test_static_spendable_outputs_justice_tx_revoked_commitment_tx() {
 	let nodes = create_network(2, &[None, None]);
 
 	// Create some initial channels
-	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	let payment_preimage = route_payment(&nodes[0], &vec!(&nodes[1])[..], 3000000).0;
 	let revoked_local_txn = nodes[0].node.channel_state.lock().unwrap().by_id.iter_mut().next().unwrap().1.channel_monitor().get_latest_local_commitment_txn();
@@ -3863,7 +3864,7 @@ fn test_static_spendable_outputs_justice_tx_revoked_htlc_timeout_tx() {
 	let nodes = create_network(2, &[None, None]);
 
 	// Create some initial channels
-	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	let payment_preimage = route_payment(&nodes[0], &vec!(&nodes[1])[..], 3000000).0;
 	let revoked_local_txn = nodes[0].node.channel_state.lock().unwrap().by_id.get_mut(&chan_1.2).unwrap().channel_monitor().get_latest_local_commitment_txn();
@@ -3907,7 +3908,7 @@ fn test_static_spendable_outputs_justice_tx_revoked_htlc_success_tx() {
 	let nodes = create_network(2, &[None, None]);
 
 	// Create some initial channels
-	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	let payment_preimage = route_payment(&nodes[0], &vec!(&nodes[1])[..], 3000000).0;
 	let revoked_local_txn = nodes[1].node.channel_state.lock().unwrap().by_id.get_mut(&chan_1.2).unwrap().channel_monitor().get_latest_local_commitment_txn();
@@ -3960,8 +3961,8 @@ fn test_onchain_to_onchain_claim() {
 	let nodes = create_network(3, &[None, None, None]);
 
 	// Create some initial channels
-	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
+	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, InitFeatures::supported(), InitFeatures::supported());
 
 	// Rebalance the network a bit by relaying one payment through all the channels ...
 	send_payment(&nodes[0], &vec!(&nodes[1], &nodes[2])[..], 8000000, 8_000_000);
@@ -4049,8 +4050,8 @@ fn test_duplicate_payment_hash_one_failure_one_success() {
 	// We route 2 payments with same hash between B and C, one will be timeout, the other successfully claim
 	let mut nodes = create_network(3, &[None, None, None]);
 
-	create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
+	let chan_2 = create_announced_chan_between_nodes(&nodes, 1, 2, InitFeatures::supported(), InitFeatures::supported());
 
 	let (our_payment_preimage, duplicate_payment_hash) = route_payment(&nodes[0], &vec!(&nodes[1], &nodes[2])[..], 900000);
 	*nodes[0].network_payment_count.borrow_mut() -= 1;
@@ -4166,7 +4167,7 @@ fn test_dynamic_spendable_outputs_local_htlc_success_tx() {
 	let nodes = create_network(2, &[None, None]);
 
 	// Create some initial channels
-	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	let payment_preimage = route_payment(&nodes[0], &vec!(&nodes[1])[..], 9000000).0;
 	let local_txn = nodes[1].node.channel_state.lock().unwrap().by_id.get_mut(&chan_1.2).unwrap().channel_monitor().get_latest_local_commitment_txn();
@@ -4214,11 +4215,11 @@ fn do_test_fail_backwards_unrevoked_remote_announce(deliver_last_raa: bool, anno
 	// And test where C fails back to A/B when D announces its latest commitment transaction
 	let nodes = create_network(6, &[None, None, None, None, None, None]);
 
-	create_announced_chan_between_nodes(&nodes, 0, 2, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	create_announced_chan_between_nodes(&nodes, 1, 2, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	let chan = create_announced_chan_between_nodes(&nodes, 2, 3, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	create_announced_chan_between_nodes(&nodes, 3, 4, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	create_announced_chan_between_nodes(&nodes, 3, 5, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	create_announced_chan_between_nodes(&nodes, 0, 2, InitFeatures::supported(), InitFeatures::supported());
+	create_announced_chan_between_nodes(&nodes, 1, 2, InitFeatures::supported(), InitFeatures::supported());
+	let chan = create_announced_chan_between_nodes(&nodes, 2, 3, InitFeatures::supported(), InitFeatures::supported());
+	create_announced_chan_between_nodes(&nodes, 3, 4, InitFeatures::supported(), InitFeatures::supported());
+	create_announced_chan_between_nodes(&nodes, 3, 5, InitFeatures::supported(), InitFeatures::supported());
 
 	// Rebalance and check output sanity...
 	send_payment(&nodes[0], &[&nodes[2], &nodes[3], &nodes[4]], 500000, 500_000);
@@ -4454,7 +4455,7 @@ fn test_dynamic_spendable_outputs_local_htlc_timeout_tx() {
 	let nodes = create_network(2, &[None, None]);
 
 	// Create some initial channels
-	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan_1 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	route_payment(&nodes[0], &vec!(&nodes[1])[..], 9000000).0;
 	let local_txn = nodes[0].node.channel_state.lock().unwrap().by_id.get_mut(&chan_1.2).unwrap().channel_monitor().get_latest_local_commitment_txn();
@@ -4488,7 +4489,7 @@ fn test_dynamic_spendable_outputs_local_htlc_timeout_tx() {
 fn test_static_output_closing_tx() {
 	let nodes = create_network(2, &[None, None]);
 
-	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	send_payment(&nodes[0], &vec!(&nodes[1])[..], 8000000, 8_000_000);
 	let closing_tx = close_channel(&nodes[0], &nodes[1], &chan.2, chan.3, true).2;
@@ -4507,7 +4508,7 @@ fn test_static_output_closing_tx() {
 
 fn do_htlc_claim_local_commitment_only(use_dust: bool) {
 	let nodes = create_network(2, &[None, None]);
-	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	let (our_payment_preimage, _) = route_payment(&nodes[0], &[&nodes[1]], if use_dust { 50000 } else { 3000000 });
 
@@ -4544,7 +4545,7 @@ fn do_htlc_claim_local_commitment_only(use_dust: bool) {
 
 fn do_htlc_claim_current_remote_commitment_only(use_dust: bool) {
 	let mut nodes = create_network(2, &[None, None]);
-	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	let route = nodes[0].router.get_route(&nodes[1].node.get_our_node_id(), None, &Vec::new(), if use_dust { 50000 } else { 3000000 }, TEST_FINAL_CLTV).unwrap();
 	let (_, payment_hash) = get_payment_preimage_hash!(nodes[0]);
@@ -4569,7 +4570,7 @@ fn do_htlc_claim_current_remote_commitment_only(use_dust: bool) {
 
 fn do_htlc_claim_previous_remote_commitment_only(use_dust: bool, check_revoke_no_close: bool) {
 	let nodes = create_network(3, &[None, None, None]);
-	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	// Fail the payment, but don't deliver A's final RAA, resulting in the HTLC only being present
 	// in B's previous (unrevoked) commitment transaction, but none of A's commitment transactions.
@@ -4855,7 +4856,7 @@ fn test_onion_failure() {
 	for node in nodes.iter() {
 		*node.keys_manager.override_session_priv.lock().unwrap() = Some(SecretKey::from_slice(&[3; 32]).unwrap());
 	}
-	let channels = [create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported()), create_announced_chan_between_nodes(&nodes, 1, 2, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported())];
+	let channels = [create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported()), create_announced_chan_between_nodes(&nodes, 1, 2, InitFeatures::supported(), InitFeatures::supported())];
 	let (_, payment_hash) = get_payment_preimage_hash!(nodes[0]);
 	let route = nodes[0].router.get_route(&nodes[2].node.get_our_node_id(), None, &Vec::new(), 40000, TEST_FINAL_CLTV).unwrap();
 	// positve case
@@ -5076,7 +5077,7 @@ fn bolt2_open_channel_sending_node_checks_part1() { //This test needs to be on i
 	let push_msat=10001;
 	nodes[0].node.create_channel(nodes[1].node.get_our_node_id(), channel_value_satoshis, push_msat, 42).unwrap();
 	let node0_to_1_send_open_channel = get_event_msg!(nodes[0], MessageSendEvent::SendOpenChannel, nodes[1].node.get_our_node_id());
-	nodes[1].node.handle_open_channel(&nodes[0].node.get_our_node_id(), Features::<FeatureContextInit>::supported(), &node0_to_1_send_open_channel);
+	nodes[1].node.handle_open_channel(&nodes[0].node.get_our_node_id(), InitFeatures::supported(), &node0_to_1_send_open_channel);
 
 	//Create a second channel with a channel_id collision
 	assert!(nodes[0].node.create_channel(nodes[0].node.get_our_node_id(), channel_value_satoshis, push_msat, 42).is_err());
@@ -5133,7 +5134,7 @@ fn test_update_add_htlc_bolt2_sender_value_below_minimum_msat() {
 	//BOLT2 Requirement: MUST offer amount_msat greater than 0.
 	//BOLT2 Requirement: MUST NOT offer amount_msat below the receiving node's htlc_minimum_msat (same validation check catches both of these)
 	let mut nodes = create_network(2, &[None, None]);
-	let _chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 100000, 95000000, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let _chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 100000, 95000000, InitFeatures::supported(), InitFeatures::supported());
 	let mut route = nodes[0].router.get_route(&nodes[1].node.get_our_node_id(), None, &[], 100000, TEST_FINAL_CLTV).unwrap();
 	let (_, our_payment_hash) = get_payment_preimage_hash!(nodes[0]);
 
@@ -5155,7 +5156,7 @@ fn test_update_add_htlc_bolt2_sender_cltv_expiry_too_high() {
 	//BOLT 2 Requirement: MUST set cltv_expiry less than 500000000.
 	//It is enforced when constructing a route.
 	let mut nodes = create_network(2, &[None, None]);
-	let _chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 100000, 0, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let _chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 100000, 0, InitFeatures::supported(), InitFeatures::supported());
 	let route = nodes[0].router.get_route(&nodes[1].node.get_our_node_id(), None, &[], 100000000, 500000001).unwrap();
 	let (_, our_payment_hash) = get_payment_preimage_hash!(nodes[0]);
 
@@ -5174,7 +5175,7 @@ fn test_update_add_htlc_bolt2_sender_exceed_max_htlc_num_and_htlc_id_increment()
 	//BOLT 2 Requirement: for the first HTLC it offers MUST set id to 0.
 	//BOLT 2 Requirement: MUST increase the value of id by 1 for each successive offer.
 	let mut nodes = create_network(2, &[None, None]);
-	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 0, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 0, InitFeatures::supported(), InitFeatures::supported());
 	let max_accepted_htlcs = nodes[1].node.channel_state.lock().unwrap().by_id.get(&chan.2).unwrap().their_max_accepted_htlcs as u64;
 
 	for i in 0..max_accepted_htlcs {
@@ -5218,7 +5219,7 @@ fn test_update_add_htlc_bolt2_sender_exceed_max_htlc_value_in_flight() {
 	//BOLT 2 Requirement: if the sum of total offered HTLCs would exceed the remote's max_htlc_value_in_flight_msat: MUST NOT add an HTLC.
 	let mut nodes = create_network(2, &[None, None]);
 	let channel_value = 100000;
-	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, channel_value, 0, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, channel_value, 0, InitFeatures::supported(), InitFeatures::supported());
 	let max_in_flight = get_channel_value_stat!(nodes[0], chan.2).their_max_htlc_value_in_flight_msat;
 
 	send_payment(&nodes[0], &vec!(&nodes[1])[..], max_in_flight, max_in_flight);
@@ -5243,7 +5244,7 @@ fn test_update_add_htlc_bolt2_sender_exceed_max_htlc_value_in_flight() {
 fn test_update_add_htlc_bolt2_receiver_check_amount_received_more_than_min() {
 	//BOLT2 Requirement: receiving an amount_msat equal to 0, OR less than its own htlc_minimum_msat -> SHOULD fail the channel.
 	let mut nodes = create_network(2, &[None, None]);
-	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 100000, 95000000, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 100000, 95000000, InitFeatures::supported(), InitFeatures::supported());
 	let htlc_minimum_msat: u64;
 	{
 		let chan_lock = nodes[0].node.channel_state.lock().unwrap();
@@ -5266,7 +5267,7 @@ fn test_update_add_htlc_bolt2_receiver_check_amount_received_more_than_min() {
 fn test_update_add_htlc_bolt2_receiver_sender_can_afford_amount_sent() {
 	//BOLT2 Requirement: receiving an amount_msat that the sending node cannot afford at the current feerate_per_kw (while maintaining its channel reserve): SHOULD fail the channel
 	let mut nodes = create_network(2, &[None, None]);
-	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 100000, 95000000, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 100000, 95000000, InitFeatures::supported(), InitFeatures::supported());
 
 	let their_channel_reserve = get_channel_value_stat!(nodes[0], chan.2).channel_reserve_msat;
 
@@ -5289,7 +5290,7 @@ fn test_update_add_htlc_bolt2_receiver_check_max_htlc_limit() {
 	//BOLT 2 Requirement: if a sending node adds more than its max_accepted_htlcs HTLCs to its local commitment transaction: SHOULD fail the channel
 	//BOLT 2 Requirement: MUST allow multiple HTLCs with the same payment_hash.
 	let mut nodes = create_network(2, &[None, None]);
-	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 100000, 95000000, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 100000, 95000000, InitFeatures::supported(), InitFeatures::supported());
 	let route = nodes[0].router.get_route(&nodes[1].node.get_our_node_id(), None, &[], 3999999, TEST_FINAL_CLTV).unwrap();
 	let (_, our_payment_hash) = get_payment_preimage_hash!(nodes[0]);
 
@@ -5330,7 +5331,7 @@ fn test_update_add_htlc_bolt2_receiver_check_max_htlc_limit() {
 fn test_update_add_htlc_bolt2_receiver_check_max_in_flight_msat() {
 	//OR adds more than its max_htlc_value_in_flight_msat worth of offered HTLCs to its local commitment transaction: SHOULD fail the channel
 	let mut nodes = create_network(2, &[None, None]);
-	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 1000000, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 1000000, InitFeatures::supported(), InitFeatures::supported());
 	let route = nodes[0].router.get_route(&nodes[1].node.get_our_node_id(), None, &[], 1000000, TEST_FINAL_CLTV).unwrap();
 	let (_, our_payment_hash) = get_payment_preimage_hash!(nodes[0]);
 	nodes[0].node.send_payment(route, our_payment_hash).unwrap();
@@ -5348,7 +5349,7 @@ fn test_update_add_htlc_bolt2_receiver_check_max_in_flight_msat() {
 fn test_update_add_htlc_bolt2_receiver_check_cltv_expiry() {
 	//BOLT2 Requirement: if sending node sets cltv_expiry to greater or equal to 500000000: SHOULD fail the channel.
 	let mut nodes = create_network(2, &[None, None]);
-	create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 100000, 95000000, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 100000, 95000000, InitFeatures::supported(), InitFeatures::supported());
 	let route = nodes[0].router.get_route(&nodes[1].node.get_our_node_id(), None, &[], 3999999, TEST_FINAL_CLTV).unwrap();
 	let (_, our_payment_hash) = get_payment_preimage_hash!(nodes[0]);
 	nodes[0].node.send_payment(route, our_payment_hash).unwrap();
@@ -5368,7 +5369,7 @@ fn test_update_add_htlc_bolt2_receiver_check_repeated_id_ignore() {
 	// We test this by first testing that that repeated HTLCs pass commitment signature checks
 	// after disconnect and that non-sequential htlc_ids result in a channel failure.
 	let mut nodes = create_network(2, &[None, None]);
-	create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 	let route = nodes[0].router.get_route(&nodes[1].node.get_our_node_id(), None, &[], 1000000, TEST_FINAL_CLTV).unwrap();
 	let (_, our_payment_hash) = get_payment_preimage_hash!(nodes[0]);
 	nodes[0].node.send_payment(route, our_payment_hash).unwrap();
@@ -5409,7 +5410,7 @@ fn test_update_fulfill_htlc_bolt2_update_fulfill_htlc_before_commitment() {
 	//BOLT 2 Requirement: until the corresponding HTLC is irrevocably committed in both sides' commitment transactions:	MUST NOT send an update_fulfill_htlc, update_fail_htlc, or update_fail_malformed_htlc.
 
 	let mut nodes = create_network(2, &[None, None]);
-	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	let route = nodes[0].router.get_route(&nodes[1].node.get_our_node_id(), None, &[], 1000000, TEST_FINAL_CLTV).unwrap();
 	let (our_payment_preimage, our_payment_hash) = get_payment_preimage_hash!(nodes[0]);
@@ -5436,7 +5437,7 @@ fn test_update_fulfill_htlc_bolt2_update_fail_htlc_before_commitment() {
 	//BOLT 2 Requirement: until the corresponding HTLC is irrevocably committed in both sides' commitment transactions:	MUST NOT send an update_fulfill_htlc, update_fail_htlc, or update_fail_malformed_htlc.
 
 	let mut nodes = create_network(2, &[None, None]);
-	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	let route = nodes[0].router.get_route(&nodes[1].node.get_our_node_id(), None, &[], 1000000, TEST_FINAL_CLTV).unwrap();
 	let (_, our_payment_hash) = get_payment_preimage_hash!(nodes[0]);
@@ -5463,7 +5464,7 @@ fn test_update_fulfill_htlc_bolt2_update_fail_malformed_htlc_before_commitment()
 	//BOLT 2 Requirement: until the corresponding HTLC is irrevocably committed in both sides' commitment transactions:	MUST NOT send an update_fulfill_htlc, update_fail_htlc, or update_fail_malformed_htlc.
 
 	let mut nodes = create_network(2, &[None, None]);
-	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	let route = nodes[0].router.get_route(&nodes[1].node.get_our_node_id(), None, &[], 1000000, TEST_FINAL_CLTV).unwrap();
 	let (_, our_payment_hash) = get_payment_preimage_hash!(nodes[0]);
@@ -5491,7 +5492,7 @@ fn test_update_fulfill_htlc_bolt2_incorrect_htlc_id() {
 	//BOLT 2 Requirement: A receiving node:	if the id does not correspond to an HTLC in its current commitment transaction MUST fail the channel.
 
 	let nodes = create_network(2, &[None, None]);
-	create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	let our_payment_preimage = route_payment(&nodes[0], &[&nodes[1]], 100000).0;
 
@@ -5528,7 +5529,7 @@ fn test_update_fulfill_htlc_bolt2_wrong_preimage() {
 	//BOLT 2 Requirement: A receiving node:	if the payment_preimage value in update_fulfill_htlc doesn't SHA256 hash to the corresponding HTLC payment_hash	MUST fail the channel.
 
 	let nodes = create_network(2, &[None, None]);
-	create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	let our_payment_preimage = route_payment(&nodes[0], &[&nodes[1]], 100000).0;
 
@@ -5566,7 +5567,7 @@ fn test_update_fulfill_htlc_bolt2_missing_badonion_bit_for_malformed_htlc_messag
 	//BOLT 2 Requirement: A receiving node: if the BADONION bit in failure_code is not set for update_fail_malformed_htlc MUST fail the channel.
 
 	let mut nodes = create_network(2, &[None, None]);
-	create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 1000000, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 1000000, InitFeatures::supported(), InitFeatures::supported());
 	let route = nodes[0].router.get_route(&nodes[1].node.get_our_node_id(), None, &[], 1000000, TEST_FINAL_CLTV).unwrap();
 	let (_, our_payment_hash) = get_payment_preimage_hash!(nodes[0]);
 	nodes[0].node.send_payment(route, our_payment_hash).unwrap();
@@ -5608,8 +5609,8 @@ fn test_update_fulfill_htlc_bolt2_after_malformed_htlc_message_must_forward_upda
 	//    * MUST return an error in the update_fail_htlc sent to the link which originally sent the HTLC, using the failure_code given and setting the data to sha256_of_onion.
 
 	let mut nodes = create_network(3, &[None, None, None]);
-	create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 1000000, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
-	create_announced_chan_between_nodes_with_value(&nodes, 1, 2, 1000000, 1000000, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 1000000, InitFeatures::supported(), InitFeatures::supported());
+	create_announced_chan_between_nodes_with_value(&nodes, 1, 2, 1000000, 1000000, InitFeatures::supported(), InitFeatures::supported());
 
 	let route = nodes[0].router.get_route(&nodes[2].node.get_our_node_id(), None, &Vec::new(), 100000, TEST_FINAL_CLTV).unwrap();
 	let (_, our_payment_hash) = get_payment_preimage_hash!(nodes[0]);
@@ -5683,7 +5684,7 @@ fn do_test_failure_delay_dust_htlc_local_commitment(announce_latest: bool) {
 	// HTLC could have been removed from lastest local commitment tx but still valid until we get remote RAA
 
 	let nodes = create_network(2, &[None, None]);
-	let chan =create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan =create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	let bs_dust_limit = nodes[1].node.channel_state.lock().unwrap().by_id.get(&chan.2).unwrap().our_dust_limit_satoshis;
 
@@ -5773,7 +5774,7 @@ fn test_no_failure_dust_htlc_local_commitment() {
 	// prone to error, we test here that a dummy transaction don't fail them.
 
 	let nodes = create_network(2, &[None, None]);
-	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	// Rebalance a bit
 	send_payment(&nodes[0], &vec!(&nodes[1])[..], 8000000, 8_000_000);
@@ -5827,7 +5828,7 @@ fn do_test_sweep_outbound_htlc_failure_update(revoked: bool, local: bool) {
 	// Broadcast of HTLC-timeout tx on local commitment tx, trigger failure-update of non-dust HTLCs
 
 	let nodes = create_network(3, &[None, None, None]);
-	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	let bs_dust_limit = nodes[1].node.channel_state.lock().unwrap().by_id.get(&chan.2).unwrap().our_dust_limit_satoshis;
 
@@ -5960,7 +5961,7 @@ fn test_upfront_shutdown_script() {
 	let nodes = create_network(3, &cfgs);
 
 	// We test that in case of peer committing upfront to a script, if it changes at closing, we refuse to sign
-	let flags = Features::<FeatureContextInit>::supported();
+	let flags = InitFeatures::supported();
 	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 2, 1000000, 1000000, flags.clone(), flags.clone());
 	nodes[0].node.close_channel(&OutPoint::new(chan.3.txid(), 0).to_channel_id()).unwrap();
 	let mut node_0_shutdown = get_event_msg!(nodes[0], MessageSendEvent::SendShutdown, nodes[2].node.get_our_node_id());
@@ -5996,7 +5997,7 @@ fn test_upfront_shutdown_script() {
 	}
 
 	// We test that if case of peer non-signaling we don't enforce committed script at channel opening
-	let mut flags_no = Features::<FeatureContextInit>::supported();
+	let mut flags_no = InitFeatures::supported();
 	flags_no.unset_upfront_shutdown_script();
 	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 1000000, flags_no, flags.clone());
 	nodes[0].node.close_channel(&OutPoint::new(chan.3.txid(), 0).to_channel_id()).unwrap();
@@ -6067,7 +6068,7 @@ fn test_user_configurable_csv_delay() {
 	nodes[1].node.create_channel(nodes[0].node.get_our_node_id(), 1000000, 1000000, 42).unwrap();
 	let mut open_channel = get_event_msg!(nodes[1], MessageSendEvent::SendOpenChannel, nodes[0].node.get_our_node_id());
 	open_channel.to_self_delay = 200;
-	if let Err(error) = Channel::new_from_req(&test_utils::TestFeeEstimator { sat_per_kw: 253 }, &keys_manager, nodes[1].node.get_our_node_id(), Features::<FeatureContextInit>::supported(), &open_channel, 0, Arc::new(test_utils::TestLogger::new()), &low_our_to_self_config) {
+	if let Err(error) = Channel::new_from_req(&test_utils::TestFeeEstimator { sat_per_kw: 253 }, &keys_manager, nodes[1].node.get_our_node_id(), InitFeatures::supported(), &open_channel, 0, Arc::new(test_utils::TestLogger::new()), &low_our_to_self_config) {
 		match error {
 			ChannelError::Close(err) => { assert_eq!(err, "Configured with an unreasonable our_to_self_delay putting user funds at risks"); },
 			_ => panic!("Unexpected event"),
@@ -6076,10 +6077,10 @@ fn test_user_configurable_csv_delay() {
 
 	// We test msg.to_self_delay <= config.their_to_self_delay is enforced in Chanel::accept_channel()
 	nodes[0].node.create_channel(nodes[1].node.get_our_node_id(), 1000000, 1000000, 42).unwrap();
-	nodes[1].node.handle_open_channel(&nodes[0].node.get_our_node_id(), Features::<FeatureContextInit>::supported(), &get_event_msg!(nodes[0], MessageSendEvent::SendOpenChannel, nodes[1].node.get_our_node_id()));
+	nodes[1].node.handle_open_channel(&nodes[0].node.get_our_node_id(), InitFeatures::supported(), &get_event_msg!(nodes[0], MessageSendEvent::SendOpenChannel, nodes[1].node.get_our_node_id()));
 	let mut accept_channel = get_event_msg!(nodes[1], MessageSendEvent::SendAcceptChannel, nodes[0].node.get_our_node_id());
 	accept_channel.to_self_delay = 200;
-	nodes[0].node.handle_accept_channel(&nodes[1].node.get_our_node_id(), Features::<FeatureContextInit>::supported(), &accept_channel);
+	nodes[0].node.handle_accept_channel(&nodes[1].node.get_our_node_id(), InitFeatures::supported(), &accept_channel);
 	if let MessageSendEvent::HandleError { ref action, .. } = nodes[0].node.get_and_clear_pending_msg_events()[0] {
 		match action {
 			&ErrorAction::SendErrorMessage { ref msg } => {
@@ -6093,7 +6094,7 @@ fn test_user_configurable_csv_delay() {
 	nodes[1].node.create_channel(nodes[0].node.get_our_node_id(), 1000000, 1000000, 42).unwrap();
 	let mut open_channel = get_event_msg!(nodes[1], MessageSendEvent::SendOpenChannel, nodes[0].node.get_our_node_id());
 	open_channel.to_self_delay = 200;
-	if let Err(error) = Channel::new_from_req(&test_utils::TestFeeEstimator { sat_per_kw: 253 }, &keys_manager, nodes[1].node.get_our_node_id(), Features::<FeatureContextInit>::supported(), &open_channel, 0, Arc::new(test_utils::TestLogger::new()), &high_their_to_self_config) {
+	if let Err(error) = Channel::new_from_req(&test_utils::TestFeeEstimator { sat_per_kw: 253 }, &keys_manager, nodes[1].node.get_our_node_id(), InitFeatures::supported(), &open_channel, 0, Arc::new(test_utils::TestLogger::new()), &high_their_to_self_config) {
 		match error {
 			ChannelError::Close(err) => { assert_eq!(err, "They wanted our payments to be delayed by a needlessly long period"); },
 			_ => panic!("Unexpected event"),
@@ -6109,7 +6110,7 @@ fn test_data_loss_protect() {
 	// * we are able to claim our own outputs thanks to remote my_current_per_commitment_point
 	let mut nodes = create_network(2, &[None, None]);
 
-	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 1000000, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 1000000, InitFeatures::supported(), InitFeatures::supported());
 
 	// Cache node A state before any channel update
 	let previous_node_state = nodes[0].node.encode();
@@ -6226,7 +6227,7 @@ fn test_check_htlc_underpaying() {
 	let nodes = create_network(2, &[None, None, None]);
 
 	// Create some initial channels
-	create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported());
 
 	let (payment_preimage, _) = route_payment(&nodes[0], &[&nodes[1]], 10_000);
 
@@ -6271,9 +6272,9 @@ fn test_announce_disable_channels() {
 
 	let nodes = create_network(2, &[None, None]);
 
-	let short_id_1 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported()).0.contents.short_channel_id;
-	let short_id_2 = create_announced_chan_between_nodes(&nodes, 1, 0, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported()).0.contents.short_channel_id;
-	let short_id_3 = create_announced_chan_between_nodes(&nodes, 0, 1, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported()).0.contents.short_channel_id;
+	let short_id_1 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported()).0.contents.short_channel_id;
+	let short_id_2 = create_announced_chan_between_nodes(&nodes, 1, 0, InitFeatures::supported(), InitFeatures::supported()).0.contents.short_channel_id;
+	let short_id_3 = create_announced_chan_between_nodes(&nodes, 0, 1, InitFeatures::supported(), InitFeatures::supported()).0.contents.short_channel_id;
 
 	// Disconnect peers
 	nodes[0].node.peer_disconnected(&nodes[1].node.get_our_node_id(), false);
@@ -6330,7 +6331,7 @@ fn test_bump_penalty_txn_on_revoked_commitment() {
 
 	let nodes = create_network(2, &[None, None]);
 
-	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 59000000, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 59000000, InitFeatures::supported(), InitFeatures::supported());
 	let payment_preimage = route_payment(&nodes[0], &vec!(&nodes[1])[..], 3000000).0;
 	let route = nodes[1].router.get_route(&nodes[0].node.get_our_node_id(), None, &Vec::new(), 3000000, 30).unwrap();
 	send_along_route(&nodes[1], route, &vec!(&nodes[0])[..], 3000000);
@@ -6430,7 +6431,7 @@ fn test_bump_penalty_txn_on_revoked_htlcs() {
 
 	let nodes = create_network(2, &[None, None]);
 
-	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 59000000, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 59000000, InitFeatures::supported(), InitFeatures::supported());
 	// Lock HTLC in both directions
 	let payment_preimage = route_payment(&nodes[0], &vec!(&nodes[1])[..], 3_000_000).0;
 	route_payment(&nodes[1], &vec!(&nodes[0])[..], 3_000_000).0;
@@ -6574,7 +6575,7 @@ fn test_bump_penalty_txn_on_remote_commitment() {
 
 	let nodes = create_network(2, &[None, None]);
 
-	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 59000000, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 59000000, InitFeatures::supported(), InitFeatures::supported());
 	let payment_preimage = route_payment(&nodes[0], &vec!(&nodes[1])[..], 3000000).0;
 	route_payment(&nodes[1], &vec!(&nodes[0])[..], 3000000).0;
 
@@ -6681,7 +6682,7 @@ fn test_set_outpoints_partial_claiming() {
 	// - disconnect tx, see no tx anymore
 	let nodes = create_network(2, &[None, None]);
 
-	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 59000000, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 59000000, InitFeatures::supported(), InitFeatures::supported());
 	let payment_preimage_1 = route_payment(&nodes[1], &vec!(&nodes[0])[..], 3_000_000).0;
 	let payment_preimage_2 = route_payment(&nodes[1], &vec!(&nodes[0])[..], 3_000_000).0;
 
@@ -6773,7 +6774,7 @@ fn test_bump_txn_sanitize_tracking_maps() {
 
 	let nodes = create_network(2, &[None, None]);
 
-	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 59000000, Features::<FeatureContextInit>::supported(), Features::<FeatureContextInit>::supported());
+	let chan = create_announced_chan_between_nodes_with_value(&nodes, 0, 1, 1000000, 59000000, InitFeatures::supported(), InitFeatures::supported());
 	// Lock HTLC in both directions
 	let payment_preimage = route_payment(&nodes[0], &vec!(&nodes[1])[..], 9_000_000).0;
 	route_payment(&nodes[1], &vec!(&nodes[0])[..], 9_000_000).0;
